for myself (#1)
* Fixed console.log output

In modern browsers console.log(1, 3) results in "1 3", not "1,3"

* Fix typo

* Fix some typos in Async & Performance

* fixing URLs in issue/pr templates

* typo: mistakingly -> mistakenly

* Fix broken links in Titles section of README.

Recently the links on the README to the various titles within the repository became broken. This commit replaces spaces in the subdirectory names with "\%20" which causes Github markdown to correctly display the link and for the link to properly direct to the appropriate subdirectory. (The links were working fine as of a month ago when last I checked, I assume that Github recently changed how they process links in markdown and what once worked, ceased to).